+++ src/views/BrowseInventory/BrowseInventory.js
@@ -3,6 +3,7 @@ import { useIntl } from 'react-intl';
 import { useHistory, useLocation } from 'react-router-dom';
 
 import {
+  TitleManager,
   useNamespace,
 } from '@folio/stripes/core';
 import {
@@ -46,7 +47,10 @@ const BrowseInventory = () => {
   const { isFiltersOpened, toggleFilters } = useFiltersToogle(`${namespace}/filters`);
   const { deleteItemToView } = useItemToView('browse');
   const [pageConfig, setPageConfig] = useState(getLastBrowseOffset());
+  const hasFocusedSearchOptionOnMount = useRef(false);
   const inputRef = useRef();
+  const indexRef = useRef();
+  const paneTitleRef = useRef();
 
   const { search } = location;
 
@@ -79,8 +83,17 @@ const BrowseInventory = () => {
     return filters;
   }, [filters]);
 
+  const pageTitle = useMemo(() => {
+    if (!withExtraFilters.query) {
+      return null;
+    }
+
+    return intl.formatMessage({ id: 'ui-inventory.documentTitle.browse' }, { query: withExtraFilters.query });
+  }, [withExtraFilters.query]);
+
   const {
     data,
+    isFetched,
     isFetching,
     pagination,
     totalRecords,
@@ -128,7 +141,10 @@ const BrowseInventory = () => {
   const onApplySearch = useCallback(() => {
     const isSearchQueryValid = validateDataQuery(searchQuery);
 
-    if (isSearchQueryValid) applySearch();
+    if (isSearchQueryValid) {
+      applySearch();
+      paneTitleRef.current.focus();
+    }
   }, [searchQuery, filters]);
 
   const onChangeSearchIndex = useCallback((e) => {
@@ -141,11 +157,23 @@ const BrowseInventory = () => {
     inputRef.current.focus();
   }, [inputRef.current]);
 
+  useEffect(() => {
+    if (hasFocusedSearchOptionOnMount.current) {
+      return;
+    }
+
+    if (!search || (search && isFetched && !isFetching)) {
+      indexRef.current.focus();
+      hasFocusedSearchOptionOnMount.current = true;
+    }
+  }, [isFetched, search, isFetching]);
+
   return (
     <PersistedPaneset
       appId={namespace}
       id="browse-inventory"
     >
+      <TitleManager page={pageTitle} />
       {isFiltersOpened && (
         <FiltersPane
           id="browse-inventory-filters-pane"
@@ -156,6 +184,7 @@ const BrowseInventory = () => {
           />
 
           <SingleSearchForm
+            autoFocus={false}
             applySearch={onApplySearch}
             changeSearch={changeSearch}
             disabled={!searchIndex}
@@ -167,6 +196,7 @@ const BrowseInventory = () => {
             selectedIndex={searchIndex}
             searchableIndexesPlaceholder={searchableIndexesPlaceholder}
             inputType="textarea"
+            indexRef={indexRef}
             inputRef={inputRef}
           />
 
@@ -189,6 +219,7 @@ const BrowseInventory = () => {
         isFetching={isFetching}
         isFiltersOpened={isFiltersOpened}
         pagination={pagination}
+        paneTitleRef={paneTitleRef}
         toggleFiltersPane={toggleFilters}
         totalRecords={totalRecords}
       />